requirements.txt
pandas
numpy
scikit-learn
sqlalchemy
psycopg2-binary
mlflow
pyyaml
joblib
tqdm